Move DataContainer voters to correct namespace
